# Signficantly improve inhibitor performance via new cache datastructure