@Spaceman1701 2025-04-07
I just updated the PR to include a new benchmark that's the same as `1_inhibition_rule,_X_inhibiting_alerts` that uses the `Equal` field so that the target labelset is only inhibit by the last source alert. I think this isolates the O(1) behavior of the new implementation pretty well.

Here's a quick 1 iteration run on my machine:
```
$ benchstat 1-run-main.txt 1-run-patch.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/alertmanager/inhibit
cpu: AMD EPYC Processor (with IBPB)
                                                                   │ 1-run-main.txt │            1-run-patch.txt            │
                                                                   │     sec/op     │    sec/op     vs base                 │
Mutes/1_inhibition_rule,_1_inhibiting_alert-4                          1.878µ ± ∞ ¹   1.824µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10_inhibition_rules,_1_inhibiting_alert-4                        1.893µ ± ∞ ¹   1.758µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/100_inhibition_rules,_1_inhibiting_alert-4                       1.890µ ± ∞ ¹   1.663µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1000_inhibition_rules,_1_inhibiting_alert-4                      2.108µ ± ∞ ¹   1.754µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10000_inhibition_rules,_1_inhibiting_alert-4                     1.815µ ± ∞ ¹   1.513µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10_inhibiting_alerts-4                        1.902µ ± ∞ ¹   1.511µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_100_inhibiting_alerts-4                       3.561µ ± ∞ ¹   1.505µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_1000_inhibiting_alerts-4                     23.580µ ± ∞ ¹   1.608µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10000_inhibiting_alerts-4                   177.102µ ± ∞ ¹   1.573µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10_inhibiting_alerts_w/_random_match-4        2.480µ ± ∞ ¹   2.155µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_100_inhibiting_alerts_w/_random_match-4       7.440µ ± ∞ ¹   2.228µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_1000_inhibiting_alerts_w/_random_match-4     23.019µ ± ∞ ¹   1.698µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10000_inhibiting_alerts_w/_random_match-4   176.798µ ± ∞ ¹   1.740µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/100_inhibition_rules,_1000_inhibiting_alerts-4                  19.699µ ± ∞ ¹   1.455µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10_inhibition_rules,_last_rule_matches-4                         2.834µ ± ∞ ¹   1.989µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/100_inhibition_rules,_last_rule_matches-4                       13.936µ ± ∞ ¹   7.869µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1000_inhibition_rules,_last_rule_matches-4                      123.32µ ± ∞ ¹   74.61µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10000_inhibition_rules,_last_rule_matches-4                     1238.7µ ± ∞ ¹   602.5µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                11.09µ         3.192µ        -71.23%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

I'll also setup a version that just switches `scache` to a `map[model.Fingerprint]*types.Alert` and see how it does in comparison.