@Spaceman1701 2025-04-07
Ah, I typo-d the benchmark function name for `1_inhibition_rule,_1000_inhibiting_alerts_w/_random_match-4` and `1_inhibition_rule,_10000_inhibiting_alerts_w/_random_match-4` which explains the strange scaling there.

I just pushed a fixed version and things look more logical:
```
$ benchstat 1-run-remove-store.txt 1-run-patch.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/alertmanager/inhibit
cpu: AMD EPYC Processor (with IBPB)
                                                                 │ 1-run-remove-store.txt │            1-run-patch.txt            │
                                                                 │         sec/op         │    sec/op     vs base                 │
Mutes/1_inhibition_rule,_1_inhibiting_alert-4                                1.753µ ± ∞ ¹   1.515µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10_inhibition_rules,_1_inhibiting_alert-4                              1.677µ ± ∞ ¹   1.476µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/100_inhibition_rules,_1_inhibiting_alert-4                             1.727µ ± ∞ ¹   1.425µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1000_inhibition_rules,_1_inhibiting_alert-4                            1.596µ ± ∞ ¹   1.425µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10000_inhibition_rules,_1_inhibiting_alert-4                           1.671µ ± ∞ ¹   1.435µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10_inhibiting_alerts-4                              1.882µ ± ∞ ¹   1.665µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_100_inhibiting_alerts-4                             1.881µ ± ∞ ¹   1.349µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_1000_inhibiting_alerts-4                            2.012µ ± ∞ ¹   1.404µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10000_inhibiting_alerts-4                           2.146µ ± ∞ ¹   1.534µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10_inhibiting_alerts_w/_last_match-4                2.501µ ± ∞ ¹   2.126µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_100_inhibiting_alerts_w/_last_match-4               6.389µ ± ∞ ¹   2.046µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_1000_inhibiting_alerts_w/_last_match-4             54.494µ ± ∞ ¹   2.575µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1_inhibition_rule,_10000_inhibiting_alerts_w/_last_match-4           637.438µ ± ∞ ¹   1.746µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/100_inhibition_rules,_1000_inhibiting_alerts-4                         1.696µ ± ∞ ¹   1.497µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10_inhibition_rules,_last_rule_matches-4                               2.862µ ± ∞ ¹   2.153µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/100_inhibition_rules,_last_rule_matches-4                             13.952µ ± ∞ ¹   7.199µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/1000_inhibition_rules,_last_rule_matches-4                            120.72µ ± ∞ ¹   64.23µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Mutes/10000_inhibition_rules,_last_rule_matches-4                           1191.2µ ± ∞ ¹   608.0µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                      6.855µ         3.062µ        -55.34%
```

The baseline version has `store.Alerts` replaced with a plain map, so the performance improvement is not caused by saving allocations from `scache.List()`.