@Spaceman1701 2025-04-07
Just for completeness, here's a full 6 iteration run which compares the implementation on `main` with the two local changes we discussed in the CNCF slack:
- `store.Alerts` replaced with a map
- similar `excludeSourceAndTargetMatch` logic as in this PR (so that the `SourceMatchers.Matches` call is lazy and the `TargetMatchers.Matches` call is pre-computed)

```
$ benchstat full-bench-remove-store-updated.txt full-bench-patch.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/alertmanager/inhibit
cpu: AMD EPYC Processor (with IBPB)
                                                                 │ full-bench-remove-store-updated.txt │        full-bench-patch.txt         │
                                                                 │               sec/op                │    sec/op     vs base               │
Mutes/1_inhibition_rule,_1_inhibiting_alert-4                                             1.712µ ±  6%   1.410µ ± 31%        ~ (p=0.065 n=6)
Mutes/10_inhibition_rules,_1_inhibiting_alert-4                                           1.792µ ±  5%   1.325µ ± 18%  -26.04% (p=0.002 n=6)
Mutes/100_inhibition_rules,_1_inhibiting_alert-4                                          1.707µ ± 20%   1.401µ ± 11%  -17.96% (p=0.002 n=6)
Mutes/1000_inhibition_rules,_1_inhibiting_alert-4                                         1.756µ ± 30%   1.433µ ± 29%  -18.42% (p=0.026 n=6)
Mutes/10000_inhibition_rules,_1_inhibiting_alert-4                                        1.925µ ± 53%   1.987µ ± 52%        ~ (p=0.699 n=6)
Mutes/1_inhibition_rule,_10_inhibiting_alerts-4                                           1.800µ ± 10%   1.594µ ± 55%        ~ (p=0.240 n=6)
Mutes/1_inhibition_rule,_100_inhibiting_alerts-4                                          2.232µ ± 29%   1.456µ ± 16%  -34.74% (p=0.002 n=6)
Mutes/1_inhibition_rule,_1000_inhibiting_alerts-4                                         1.724µ ± 27%   1.397µ ±  8%  -18.97% (p=0.002 n=6)
Mutes/1_inhibition_rule,_10000_inhibiting_alerts-4                                        1.991µ ± 34%   1.419µ ± 29%  -28.73% (p=0.002 n=6)
Mutes/1_inhibition_rule,_10_inhibiting_alerts_w/_last_match-4                             2.418µ ±  8%   1.782µ ±  8%  -26.31% (p=0.002 n=6)
Mutes/1_inhibition_rule,_100_inhibiting_alerts_w/_last_match-4                            6.822µ ± 13%   1.755µ ± 20%  -74.28% (p=0.002 n=6)
Mutes/1_inhibition_rule,_1000_inhibiting_alerts_w/_last_match-4                          53.252µ ±  7%   1.834µ ± 29%  -96.56% (p=0.002 n=6)
Mutes/1_inhibition_rule,_10000_inhibiting_alerts_w/_last_match-4                        635.912µ ± 24%   1.860µ ± 22%  -99.71% (p=0.002 n=6)
Mutes/100_inhibition_rules,_1000_inhibiting_alerts-4                                      1.686µ ±  6%   1.388µ ±  5%  -17.68% (p=0.002 n=6)
Mutes/10_inhibition_rules,_last_rule_matches-4                                            2.736µ ± 17%   1.986µ ±  8%  -27.40% (p=0.002 n=6)
Mutes/100_inhibition_rules,_last_rule_matches-4                                          13.307µ ±  7%   8.368µ ± 13%  -37.12% (p=0.002 n=6)
Mutes/1000_inhibition_rules,_last_rule_matches-4                                         123.30µ ±  8%   60.93µ ±  8%  -50.58% (p=0.002 n=6)
Mutes/10000_inhibition_rules,_last_rule_matches-4                                        1162.8µ ± 14%   616.1µ ±  6%  -47.02% (p=0.002 n=6)
geomean                                                                                   6.882µ         2.966µ        -56.90%
```